Move calculate_cost script to scripts folder. Keep nosetest in tests folder and point to the script in the scripts to run. #100
