+++ scripts/__init__.py
@@ -1,0 +1,2 @@
+from .calculate_cost import init_fasttrips
+from .calculate_cost import run_calculate_cost
